+++ priv/repo/migrations/20231024224850_add_blog_preferences.exs
@@ -5,6 +5,8 @@ defmodule Teiserver.Repo.Migrations.AddBlogPreferences do
     create table(:microblog_user_preferences, primary_key: false) do
       add :user_id, references(:account_users, on_delete: :nothing), primary_key: true
 
+      add :tag_mode, :string
+
       add :enabled_tags, {:array, :integer}
       add :disabled_tags, {:array, :integer}