+++ changelog.md
@@ -6,7 +6,7 @@
 - Spring `c.user.whois` command added
 - Added microblog report
 - Added RSS feed to microblog
-- ... Currently working on making tag filters work for the microblog ...
+- Tags for the microblog can now be filtered
 
 #### Bugfixes